+++ packages/ragbits-document-search/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Update audit imports (#427)
 
+## 0.17.1 (2025-05-09)
+
+### Changed
+
+- ragbits-core updated to version v0.17.1
+
 ## 0.17.0 (2025-05-06)
 
 ### Changed